+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2019-04-20
+* Fix visibility tracks not being properly keyed; previously every frame but the last one was being overwritten with new actions.
+* Implement method to return the exact name of an object - used to determine if a mesh name matches a given track name.
+* Change warning for NUANMB importer to warn about models that don't have bones aligned correctly.
+
 ## [1.3.0] - 2019-04-04
 * Stop overwriting object origins; this is needed so that everything will properly rotate around the X axis.
 * Improved method of storing vertex color and UV values, theoretically allowing for more than five of color maps or UV maps to be read.